Add specs to validate GET /moves response
We validate the response JSON against the JSON Schema that we also use
to document the API. The motivation for this is to test that the
(hand-coded) documentation matches reality and that in future we can
detect any regressions of the implementation from the docuemented
interface. It also lays the groundwork for validating incoming requests
against the schema.