Add Content-Type header to swagger params
When importing the swagger docs into an API client like Postman
it was missing the Content-Type header that is required on
requests by the API.

This would mean manually adding it each time the docs were imported/
re-imported.

This change adds a param header to each request to ensure the API
doesn't return an error by default.