Set default sort order to ID on moves
We were previously sorting moves by location title to help with
the frontend. This currently causing issues with pagination where
results may be returned across multiple pages as location is not
always guaranteed.

This change amends the sort to be by ID so the order returned during
pagination will always be the same.