generate Swagger documentation while running RSpec
This adds the Swagger documentation generator to the RSpec output. It's necessary because the swaggerize Rake task hangs when
generating swagger examples from real responses (because it's not able to run the requests against a live server.

Generating Swagger docs as part of the test run produces examples. Running the rake task leaves the examples empty