Working RSpec run by segregating new and old Swagger schemas :(.
I had been hoping to use the existing swagger definitions in-place, however this doesn't seem to be possible because of a bug in RSwag (https://github.com/rswag/rswag/issues/270)
which causes the relative paths to fail to load. Attempting to update these to internally reference /definitions/ causes the existing test base to fail
because they load the schemas from the swagger directory.

This commit does cause duplication of the code, however once all the tests are written in RSwag it will be possible to delete any legacy files and clean up