Supplier/Location rake task replaces locations for each supplier
Prior to this commit the rake task would attempt to add any locations listed in the hash to each supplier, but would not remove any already in place.

This commit alters that behaviour to make the list in the Rake task authoritative. This enables (for example) moving a location from one supplier to another with a single change/command.