Integrate Swagger into move tests (#188)
* Adds Oauth2 flow to swagger, so you can test the API directly if you have client credentials

* allow body-less requests (GET, DELETE, HEAD) to not specify a content type

* Fix /moves/ swagger doc. Uses Swagger arrays (correctly), and the appropriate response is received from the API

* Rubocop 👮‍♀️

* Bump swagger dependencies

* Switch to rswag-generated swagger doc with 0 schemas or paths

All other metadata matches the json document we're deprecating.

* Convert move show spec into RSwag format

* rename swagger (remove frontend reference), and add swagger helper setup for move show test

* generate Swagger documentation while running RSpec

This adds the Swagger documentation generator to the RSpec output. It's necessary because the swaggerize Rake task hangs when
generating swagger examples from real responses (because it's not able to run the requests against a live server.

Generating Swagger docs as part of the test run produces examples. Running the rake task leaves the examples empty

* add shared examples for 401, 404 and 415 error responses for RSwag.

RSwag carries out a bunch of validation as part of the test run, so we don't need all the shared examples currently in place. Toggling assertion
when inside an RSwag definition was tricksy, so I've duplicated the content and content-type header assertions, with a view to removing the existing shared examples if/when we've migrated to RSwag entirely

* Add schema definition to swagger content block, to populate model for swagger UI

* Automatically generate examples from RSpec responses

* rollback arbitrary whitespace addition

* Update schemas for Move (and related) to use internal address, and load through Swagger RSpec step

* Load JSON definitions to allow RSpec runner to find schemas

* latest swagger.yaml with single move endpoint

* introduce hand-coded swagger paths.

The hand-coded paths JSON contains all paths from the old swagger.json EXCEPT for GET moves/{moveId}, which
is generated by RSwag from the tests.

We should remove these definitions as we add new tests in.

* refactor NomisSync test into shared example

* rubocop 🚨

* Toggle example generation when not in RSwag so we don't break other tests

* Working RSpec run by segregating new and old Swagger schemas :(.

I had been hoping to use the existing swagger definitions in-place, however this doesn't seem to be possible because of a bug in RSwag (https://github.com/rswag/rswag/issues/270)
which causes the relative paths to fail to load. Attempting to update these to internally reference /definitions/ causes the existing test base to fail
because they load the schemas from the swagger directory.

This commit does cause duplication of the code, however once all the tests are written in RSwag it will be possible to delete any legacy files and clean up

* remove unnecessary error codes from move responses json

* Disable automatic generation of response examples (to prevent unique changes to swagger.yaml each time the tests are run)

* Refactor authorisation header to make it dry

* Use snake-case for Swagger parameters