Refactor Circle CI build to introduce explicit job for API docs Previously API documentation using swagger was generated at the end of the test job - this meant a dependency on the build job for the test job to complete, in turn slowing down the overall build pipeline. For the staging environment the test and build jobs were performed in parallel, but that introduced a race condition with the swagger file being unavailable when the docker image is created.
This approach adds a new job to generate the swagger docs which can be added as a dependency for the build job - within this job we still need to install all of the dev and test gems, but don't need to run the full test suite before making the `swagger.yml` file available.

All of the environments now include the new api_docs job as a dependency of the build jobs (so that swagger is included). The test job is then run in parallel with the build job. This should mean that overall the build and deploy for each environment is faster, and also respects the dependencies for swagger so that API documentation is generated correctly.