P4-1385 Allow move creation and serialisation without a person
jira-ticket: https://dsdmoj.atlassian.net/browse/P4-1385

When a PMU allocation is created, we would like to assocaite
them with moves, for each of the people required for the allocation.
However at the time moves are created, they will not be associated
with a person, as they will be added later by the OCA.

Allow a move to be created without a person, only if it is a request
or a cancellation. As proposed moves coming from the singleton flow
always will require a person associated, while moves created for an
association wil always start with the status request.

Also validate on uniqueness of date unless the person is not assocaited
with a move to not group nil moves together created for an allocation as
invalid.