Removes unused supplier serializer code
This [pr](https://github.com/ministryofjustice/hmpps-book-secure-move-api/pull/143/)
incorrectly added a suppliers relationship as one of the attributes of
the LocationSerializer.

This code never returns SupplierSerializer objects and actually returns
the result of just calling Location#suppliers so I can't see any harm
removing it.

If suppliers/frontend want to have this returned with a location then we
should implement this in a compliant way.