P4-1376 Move controller logic to service class
jira-ticket: https://dsdmoj.atlassian.net/browse/P4-1376

* Move creation logic for allocation out of controller into
service class. Followed conventions regarding creator service classes
same as people.
* Moved notification into after action in controller, and added tests
for not creating notifications if there are errors
* Changed logic around creating moves, where now we check for date
rather than validity of allocation. This is because we do not want to
attempt to create a move if an allocation is invalid. We know that
location will fail if it does not exist, before getting to the move
creation, so there is no risk around returning the wrong validation
message. So we are only checking for date.