Add importer class to populate regions reference data Added new region model and corresponding migration including join table with locations (regions have many locations and locations can be part of many regions) Added YML file containing details of the regions (key, name and NOMIS location id's) to populate data (reworked from original Excel spreadsheet) Added rake tasks to call importer with the YML file
The reference data for this is mastered in the back end as the source of truth. Working with a YML file is much simpler and avoids the need to add extra gems (e.g. Roo) to process spreadsheets. We can modify the YML file as needed in future should the regions change.
Future work will expose a reference data endpoint to return the locations associated with each region
This is needed for the front end allocations work to allow user to select region of interest