P4-1398 Surface moves in allocation GET endpoint
jira-ticket: https://dsdmoj.atlassian.net/browse/P4-1398

Surface moves and people associated with them in the
allocation endpoint. We are currently including a circular
dependency in move relationships for an allocation due
to the `belongs_to` in the moves serializer. Flagging this as
something to look into in further tickets.