P4-1398 Ignore moves order in API response for allocations
jira-ticket: https://dsdmoj.atlassian.net/browse/P4-1398

When returning the moves in an allocation, we should not care about
the order of the array. Unfortunately there isn't a way to compare this
without either comparing the keys in question or defining those as
`UnorderedArray`. I've gone with the latter to avoid adding tests
for each nested structure.