build: Add slack notifications to CI workflow
This change copies across the notifications from the frontend workflow
to notify the team slack channel of releases being prepared, needing
deployment approval and when they have been successfully deployed.

It also changes the deployment of pre-production to only be done for
releases so that it will more closely mimic production rather than
always containing the latest master changes.