P4-1441 Add allocation events create path
jira-ticket: https://dsdmoj.atlassian.net/browse/P4-1441

Add a new "cancel" event for allocations. When posting to this endpoint
we expect the allocation and all associated moves to be cancelled. This is
following the move event creation pattern, but is still WIP until event is
flushed out in further work. For now we return a fake event response.

To save update associations we have to set autosave to true on the
`has_many` association on the model.

Since cancellation reason is yet to be implemented by the FE, we will disable
the validations for now until it is implemented