Allocation `moves_count` now includes only non cancelled moves When the allocation is first created, the `moves_count` attribute is used to create a number of associated moves with `requested` status, so the `moves_count` attribute on the allocation is correct at that point. Subsequently, the allocation cannot be changed but may be cancelled. Additionally the associated moves may be changed - notably to amend their status.
If the allocation is cancelled completely, the `moves_count` is now correctly set to zero. This also required changing the validation on `moves_count` to only apply on create.

If a move associated to an allocation is updated (via the `Moves::Updater` service) then the allocation `moves_count` is now updated to count only non cancelled moves.

Also refactored a few specs to avoid use of the confusing `:skip_before` tag.