P4-1555 Allow nil allocations to be updated on allocations
jira-ticket: https://dsdmoj.atlassian.net/browse/P4-1555

Until we remove nil values from statuses add none trait
for allocation, which updates before a create to bypass the callbacks.
We also need to allow unfill to be called on itself, this is because
when we create a record with nil status, its then initialised with
unfill status which then we might call unfill again. We'll allow
this for now until we remove nil values.