Add new endpoint for cancelling an allocation with reason and comment New endpoint allows a POST to /allocations/:id/cancel and accepts a cancellation_reason and optional cancellation_reason_comment attribute.
The existing "events" endpoint for allocations has now been deprecated as we've decided to move towards explicit "verb" endpoints for dedicated actions. For now the existing event endpoint will be retained as this is used in the front end - but it doesn't support a custom cancellation reason or cancellation reason comment. Once the new endpoint is adopted within the front end then the deprecated events endpoint can be removed which will clean up quite a bit of code; left lots of TODO's to annotate where this needs to happen.