Support `person` in include validator
Active Model Serializers (AMS) automagically returns a root object
in the following `.` (dot) separated format: `person.gender`.

This meant that we never needed to add `person` to our static include
list.

The validator we added does not support root objects implicitly, either.

Since both the validator and AMS depend on the same list this means we
need to make sure we're supporting the explicit form for included
documents when doing validations.