P4-1582 Allow a profile to be set on move creation
jira-ticket: https://dsdmoj.atlassian.net/browse/P4-1582

Allow a move to be created by either passing in a profile id or person id.
The profile will take precedence unless a person id is passed in,
where the latest profile from this person will be set to the move.
The person will be removed after migration is complete.