P4-1582 Allow a profile to be updated on a move
jira-ticket: https://dsdmoj.atlassian.net/browse/P4-1582

Allow a move profile relationship to be updated either by
passing in a profile or a person id. If an empty profile is
passed the current profile should be unlinked. The person
implementation will be removed after migration is complete. Also
update the docs to show profiles are now returned and accepted