Reinstate support for null `move_agreed` attribute on move model This was requested by front end guys, and reverts a previous migration to set `false` as the default value for `move_agreed` if this was not specified or passed as `null` when creating a move. If omitted or passed as `null` then a `null` value gets stored.
Also added a data maintenance rake task to correct retrospective data - this is to be run once this change has been deployed and is implemented within the front end (exact timing to be finalised). This sets `move_agreed` to `null` for all non prison transfer moves where the `move_agreed` has not already been explicitly set to true.