Add support for automatically rebooking a move when rejecting Adds an additional `rebook` parameter to the move reject request Adds method to move model to rebook (create a new move) with pertinent details from original move Added new `original_move` relationship to allow rebooked moves to relate to the original move Included `original_move` in Move serializer
This allows the front end to specify that a move should be automatically be rebooked in 7 days; the new move is a proposed move with allocation, profile, location and date details copied across from the original move. This provides a convenience for the PMU/OCA to create a new move in future.