[P4-1500] Dereference Move#documents (#615)
* Dereference Move#documents

Jira link: https://dsdmoj.atlassian.net/browse/P4-1500

We've migrated away from having documents on a move to a polymorphic
version of documents on a profile. This commit removes all reference to
a document on a move.

The approach was to remove the relationship and overridden
Move#documents method and fix broken specs.