P4-1749 Add section parser for importer with framework fixtures
jira-ticket: https://dsdmoj.atlassian.net/browse/P4-1749

* Add fixtures for frameworks with two versions of PER
* Add section parser which goes through each section file and
sets initial set of questions with keys and section attributes.
It also sets section dependencies i.e. if a section is dependent
on a question value. If a section is dependent on a question all
questions under that section should have that question as a parent