P4-1750 Import flags from framework questions
jira-ticket: https://dsdmoj.atlassian.net/browse/P4-1750

Import flags for a question if a question is answered in a
certain way. A flag will have a type, label and the value it
relates to in the question.

Add an validation on framework questions on types now we have a
list of values we expect for them.