Fix importing offenders from Nomis in GET /api/people v2
Jira link: https://dsdmoj.atlassian.net/browse/P4-1852

The service we wrote and were mocking out in the controller for
importing a person from Nomis was built to operate against a single
person which is inconsistent with filter param possibilities described
by the json:api spec.

This enables us to import more than one prisoner at a time in version 2
of this endpoint.