P4-1791 Add responded field on framework responses
jira-ticket: https://dsdmoj.atlassian.net/browse/P4-1791

Add field to determine if a response has been seen/answered. We
need this to determine if someone has gone through the question
in the PER even if it is optional and not answered.

Set responded to true if value is set before saving. Using
value_text and value_json to avoid errors in PER builder