Clear dependent response values as well as flags
If the parent answer is changed, we would like to clear
all answered dependent values and flags. We should clear
the full tree of dependents and not only the first level down.
To determine the correct dependent responses to clear, only
select the responses with dependent values used to match
the answered value, but don't match it anymore (this is found
in the framework question attached to the response under
`dependent_value`).

To retrieve the full tree of descendents utilize the recursive
postgres function which will select the node and all it's
descendents and return their ids.