P4-1792 Fix progress section for the PER
jira-ticket: https://dsdmoj.atlassian.net/browse/P4-1792

The previous implementation of the progress section included
all responses and only looked at a response having responded
true or false. This is an issue with dependent questions,
where they are valid remaining unanswered if a parent
question is answered differently and doesn't lead to
them, meaning they will always remain unanswered.

Add a filter to only select and check nodes which
are not dependent responses, or dependent responses
where the parent has been answered with the correct
dependent response to include them