Add state machine to person escort record model
* Include new states unstarted and printed
* Add validations for confirmed and printed timestamps
when state is set to both
* Add method to update state on person escort record,
which utilises same query for sections, but does not group
them to the specified section to calculate the general state.
If person escort record is confirmed/printed, this method will
throw an error to stop the ability to calculate the state, as
this is no longer editable and can only be done by the patch endpoint.

* Since we can't use not_started due to enum warnings in the model,
surface this value in the serializer instead.