P4-1591 Add Person escort record PATCH endpoint
jira-ticket: https://dsdmoj.atlassian.net/browse/P4-1783

Add new PATCH endpoint for person escort record, which accepts
a status, which could be confirmed or printed, and updates the
person escort record state. It also sets the timestamps for each
respectively.

If either the start state is invalid, or any other params are passed
in return an error.