Remove dead nomis_event_ids attribute from moves It seems that in the distant past, Nomis events were imported from Nomis and used to create associated moves in PECS. When first implemented, the original Nomis event id was stored in the move as `nomis_event_id` attribute as these have a 1:1 mapping with moves. Also these imported moves were treated as read only so that they could not be modified via the API (which would presumably be a bad thing).
Time passed...

Subsequently it was discovered that court hearings break the original 1:1 mapping approach with the same prisoner potentially having multiple court hearings (for different offences) on the same date and location, which means a move cannot be created for each event. Nomis stores each individual court hearing as a separate event; so at that point `nomis_event_ids` was added to store an array of all associated Nomis events for the move. A rake task was added to back fill this array attribute on existing data, and the `nomis_event_ids` array was then used to determine if a move originated from Nomis.

Time passed...

At some point it was further decided not to import events from Nomis to create moves any longer. This meant that in production a number of unwanted moves were left that needed to be cancelled. Another rake task was built for this. There is no remaining evidence of the Nomis event importer in the code base so presumably this was removed, although the legacy `nomis_event_id` and `nomis_event_ids` plus checks for enforcing read only access to moves from Nomis was left behind. This looks to be an oversight as there is no reference to these attributes in the code which means they are not exposed via the API.

In an attempt to clean up the residual mess I have removed the `nomis_event_ids` attribute. There will be ancient cancelled moves in production that were from Nomis originally that will hence lose the association with the original Nomis event - however this data is unwanted and probably should have been deleted completely instead. The original `nomis_event_id` attribute however does have some future use, as it will be repurposed to store the created prison to prison transfer event id. This means that the custom accessor code for this attribute can be removed and it can revert to a simple integer type.