P4-1793 Add `ready_for_transit` filter on moves
jira-ticket: https://dsdmoj.atlassian.net/browse/P4-1793

Add new filter to return all moves ready to be in transit if set
to true. Moves ready for transit are moves where the person escort
record is confirmed.

A move is not ready to be in transit if there is no person escort
record attached, or if the status is not confirmed. This filter
will be used on the FE dashboard for moves.