P4-2004 Add new collection type multiple items
jira-ticket: https://dsdmoj.atlassian.net/browse/P4-2004

Add a new collection response type: multiple items. Each multiple items
list contains a multiple item object which encapsulates the item
and the corresponding responses. It applies on the validations to that
object and bubbles them up to the collection appropriatly.

This type of question has nested questions which do not have corresponding
responses, but rather those responses are encapsulated in the parent question,
mapped to the item they are associated to.

Since the responses provided for the item mirror framework responses,
validate those on update (if normal validation added it allows
them to be empty which we don't want). Scope validations to an index of the
responses and the items to avoid overwriting validation messages.