Different endpoint for returning moves in CSV format The original endpoint chosen for this aligned with Rails idioms with a GET verb - primarily chosen as it's returning a list of moves. Whilst this works fine the use of GET imposes a limitation in the length of the passed URL query string - tested to be around 8k characters. This limit is set within the web infrastructure so is not easy to get around.
Some of the users have a large list of locations they can access, so building a request that specifies each of the required location id filters results in a 10k request - this was blowing up with an error because of the request size. This could be better architected by performing location authorisation in the API rather than the front end (passing some user info to the API instead), but that's a longer term goal.

To work around the immediate problem, this PR swaps to a POST request which not impose the same size limit, but does mean that the filter params must be passed in JSON format rather than as part of the query string. The new endpoint supports the same filter parameters as `GET /moves`. We can't use `POST /moves` for this as that is used to create a new move, so we have chosen `POST /moves/csv` as the new endpoint. This is intentionally undocumented as it's not intended for supplier use (only by the front end). All references to"csv" have been removed in the swagger documentation.