P4-2061 Improve type validation on framework responses
jira-ticket: https://dsdmoj.atlassian.net/browse/P4-2061

Check the response type on setting the value in the base
response class to avoid having to check errors in validations
later on, which has caused us to duplicate type validation across
multiple points when creating/updating records to avoid random errors.

Now we first call super on value setting which checks the type, and
fails if it is invalid.