P4-2080 Add Framework response types to interface
jira-ticket: https://dsdmoj.atlassian.net/browse/P4-2080

We currently support `collection::add_multiple_items` for
a response that is a collection and of type multiple items. To
extend this pattern to other exact collections and objects, append
`::details` to object and collection if it has followup comments. If
another type of collection or object is added, we will add that type to
those respectively.