Revised business rules for determining move supplier This largely reverts previous work to introduce supplier determination when the move is updated; instead the supplier for a move is only determined when the move is initially created. Once created the assigned supplier owner never changes. Also this reverts the previous reliance solely on the SupplierLocation mapping and reinstates checks for supplier API keys. If a supplier calls the API then the move is assigned to the supplier irrespective of SupplierLocation mapping. If the front end calls the supplier then SupplierLocation mapping is used to determine the correct supplier.
Removed the previous `determine_supplier` method on move model as we don't need to check for move date or date_from changing now. Amended the params for SupplierChooser to accept either a move or allocation instance; this makes it easier to call the service and for the service to determine whether to use `date` or `date_from` accordingly.

Remove previous additions and specs to ensure that supplier is amended when patching or accepting a move. Updated request specs to check that supplier API key is used in preference to the doorkeeper service when present.

Also removed redundant `moves:update_suppliers` service. This was needed for go live, but is actually now dangerous if we were to run this again as it does not consider supplier API keys and will in fact revert any existing court moves to have a nil supplier (which we don't want!).