Add new endpoint for bulk update of PER framework responses This allows multiple framework responses to be patched at once; this would be the case for a supplier sending the entire PER responses as a single request. This leverages the bulk extension of JSONAPI spec but does not formally require any additional headers to be passed (JSONAPI extensions as a whole are still under review, so opted to keep this as simple as possible for suppliers to implement).
Refactored existing `FrameworkResponse` model to allow various actions to be performed to multiple responses at once - this largely means promoting instance methods to class methods. Introduced a new `BulkUpdater` service as this would have bloated the `FrameworkResponse` model too much. Makes extensive use of the activerecord-import gem for bulk upserts of modified data - this keeps things snappy with the entire response for updating every question in a PER at around 900ms. If not every question value is changed this drops to around 200ms which should be fast enough and is preferable to resorting to a more complex asynchronous implementation.