P4-2198 Fix unique profile when creating the PER error
jira-ticket: https://dsdmoj.atlassian.net/browse/P4-2198

When a user is double clicking create the PER button, they are seeing
a 500 error as opposed to be gracefully redirected to the already created PER.
Ideally the API should be catching this uniqueness validation error, and returning
a 422 with the relevant message, however due to the way our app is setup in production,
where we have multiple instances attempting the same thing,  this isn't being caught as
a validation error on uniqueness, and is falling back to the database level of
errors, and raises an exception.

We have to manually catch this error if it happens. I've added tests
for both instances, so now we can handle the uniqueness validation as well as
the DB error itself.