Handle event-specific relationships in generic events controller
In our api we require named relationships to be passed in the generic
events controller. This is a hangover from the original event
implementation that expects relationships to be passed in a
json:api-compliant way.

In the new world the event api expects event-specific attributes
(ignoring that they're relationships) to be in the details blob
which is optimised for reporting.

This change makes the generic api accept both ways of passing relationships.