P4-2067 Add Move to Person escort record
jira-ticket: https://dsdmoj.atlassian.net/browse/P4-2067

To allow the person escort record to access move attributes like
location, we need to know which move is associated to a person escort
record. We cannot do that through the profile, as it can be associated
to multiple moves. By associating the PER directly to a move, we can
access the move required. Since this is a transitional process, allow
the move to be optional for now.