P4-2067 Allow move to be passed in on PER creation
jira-ticket: https://dsdmoj.atlassian.net/browse/P4-2067

On PER creation, allow either a profile OR a move to be passed in.
This is temporary as this is a transition to completely utilise
moves instead of profiles. To avoid any breaking changes for API
clients we will allow both for now.

If a move is passed in, the PER is associated to the Move and the
profile is extracted from the Move.
If a profile is passed in, we use that profile and set the Move to
nil.

Expose the move relationship in the API response as well.