Switch Person escort record to accept a move only
We currently support both profile or move for creating a PER. Switch
to only use the move to create a profile, as using a profile
is deprecated, we will be associating the profile attached to the
move instead. The move is required to access the location and other
data required for creating the PER, which isn't available by just
passing the profile.