IM-60 Retry long transactions if they fail
jira-ticket: https://dsdmoj.atlassian.net/browse/IM-60

When deploying migrations, the database schema becomes invalid.
If there were transactions midway during such a deploy a
`ActiveRecord::PreparedStatementCacheExpired` error is thrown.
To avoid those errors on deploy, wrap all long running transacions
with a custom retriable transaction, which retries once if this error
is raised, but fails if it fails another time. All transactions to be
retried are safe, but are long running so are prone to raise such
errors on deploy.