Add capability of running every feed on a specific date
Enables running each report feed for the hub and jpc on a specific date.

We need this because of a reporting error in which migrations created
updates to every profile in the profiles table in production. We need to
rerun the report feeds that failed to generate for the relevant date.

This change enables specifying the REPORT_ON_DATE for all or for each of
the feeds