jira link: https://dsdmoj.atlassian.net/browse/P4-2422
The goal of this commit is to enable surfacing generic event relationships that
that are currently stored in the details attribute to be in the
relationship section.

In order to achieve that we have to first know the type of each
relationship so will add type information to the relationship_attributes
call.

We also need a method that takes all of that information for the current
event STI instance and pulls out the relationships that we're
interested in for consumption by the frontend later.

- [x] Updates GenericEvent#relationship_attributes dsl to also take a
  type
- [x] Updates all GenericEvent STI classes to pass their relationship
  type where necessary
- [x] Adds GenericEvent#relationships method to render relationship
  information
- [x] Updates all tests to reflect change to relationship attributes
- [x] Removes some now-redundant relationship methods which were used to
  retrieve the relationship